requirements.txt
fastapi
pydantic
pymupdf
python-dotenv
requests
sqlalchemy
